company/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dataclasses==0.8
defusedxml==0.7.1
Django==3.2.3
django-filter==2.4.0
django-templated-mail==1.1.1
django-utils==0.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-nested-routers==0.93.3
drf-yasg==1.20.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
